refactor: extract analysis function
